<Refactor code so pool queries are not in as many files.  Cleaned up comments>
