--- db/query.sql
@@ -1,1 +1,0 @@
-SELECT * FROM department;
