+++ package.json
@@ -4,7 +4,8 @@
     "description": "Command-line application to manage an employee database using Node.js, Inquirer, and PostgreSQL.",
     "main": "index.js",
     "scripts": {
-      "test": "echo \"Error: no test specified\" && exit 1"
+      "test": "echo \"Error: no test specified\" && exit 1",
+      "start": "node index.js"
     },
     "author": "Brian Wach",
     "license": "MIT",
